Pulling the retrieval of actions up higher for better performance
