+++ app/src/ui/toolbar/branch-dropdown.tsx
@@ -254,10 +254,6 @@ export class BranchDropdown extends React.Component<
 
   private renderPopover() {
     const pr = this.props.currentPullRequest
-    const { tip } = this.props.repositoryState.branchesState
-    // It is ok if it doesn't exist, we just can't retrieve actions workflows
-    const currentBranchName = tip.kind === TipState.Valid ? tip.branch.name : ''
-
     if (pr === null) {
       return null
     }
@@ -273,7 +269,6 @@ export class BranchDropdown extends React.Component<
               prNumber={pr.pullRequestNumber}
               dispatcher={this.props.dispatcher}
               repository={pr.base.gitHubRepository}
-              branchName={currentBranchName}
             />
           </div>
         </Popover>
@@ -283,6 +278,9 @@ export class BranchDropdown extends React.Component<
 
   private renderPullRequestInfo() {
     const pr = this.props.currentPullRequest
+    const { tip } = this.props.repositoryState.branchesState
+    // It is ok if it doesn't exist, we just can't retrieve actions workflows
+    const currentBranchName = tip.kind === TipState.Valid ? tip.branch.name : ''
 
     if (pr === null) {
       return null
@@ -294,6 +292,7 @@ export class BranchDropdown extends React.Component<
         dispatcher={this.props.dispatcher}
         repository={pr.base.gitHubRepository}
         onBadgeClick={this.onBadgeClick}
+        branchName={currentBranchName}
       />
     )
   }
